Code format (clang-format chromium style)

+++ games/zobristgenerator.cpp
@@ -1,15 +1,17 @@
 #include "zobristgenerator.h"
 
-ZobristGenerator &ZobristGenerator::GetInstance() {
+ZobristGenerator& ZobristGenerator::GetInstance() {
   static ZobristGenerator instance;
   return instance;
 }
 
-void ZobristGenerator::GenerateUniqueZobristValues(int num, std::vector<ZobristValue> &values) {
+void ZobristGenerator::GenerateUniqueZobristValues(
+    unsigned int num,
+    std::vector<ZobristValue>& values) {
   values.clear();
   values.reserve(num);
 
-  for (int i = 0; i < num; ++i) {
+  for (size_t i = 0; i < num; ++i) {
     ZobristValue value;
     do {
       value = zobrist_distribution_(random_engine_);
@@ -20,5 +22,6 @@ void ZobristGenerator::GenerateUniqueZobristValues(int num, std::vector<ZobristV
 }
 
 ZobristGenerator::ZobristGenerator()
-    : random_engine_(random_device_())
-    , zobrist_distribution_(std::numeric_limits<ZobristValue>::min(), std::numeric_limits<ZobristValue>::max()) {}
+    : random_engine_(random_device_()),
+      zobrist_distribution_(std::numeric_limits<ZobristValue>::min(),
+                            std::numeric_limits<ZobristValue>::max()) {}
